requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
decorator==4.4.2
Django==5.1.1
django-adminlte2==0.4.3
django-distill==3.2.7
django-extensions==3.2.3
django-grappelli==4.0.1
django-jet==1.0.8
django-suit==0.2.28
djangorestframework==3.15.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
incremental==24.7.2
moviepy==1.0.3
msgpack==1.1.0
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proglog==0.1.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.1.0
sqlparse==0.5.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
zope.interface==7.1.1